Revert "feat: init chapter 3 (#17)"
This reverts commit 96b1150611f738510bf8eb4e0daf922c38493034.
