--- src/chapter-2/index-5.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>3js-cpl121</title>
-    <link rel="stylesheet" href="./style.css">
-</head>
-<body>
-    <canvas class="webgl"></canvas>
-
-    <section class="section">
-        <h1>My Portfolio</h1>
-    </section>
-    <section class="section">
-        <h2>My projects</h2>
-    </section>
-    <section class="section">
-        <h2>Contact me</h2>
-    </section>
-
-    <script type="module" src="./script.js"></script>
-</body>
-</html>
