+++ src/style.css
@@ -4,11 +4,15 @@
     padding: 0;
 }
 
-html,
+html
+{
+    background: #1e1a20;
+}
+/* html,
 body
 {
     overflow: hidden;
-}
+} */
 
 .webgl
 {
@@ -16,4 +20,24 @@ body
     top: 0;
     left: 0;
     outline: none;
+}
+
+
+.section
+{
+    display: flex;
+    align-items: center;
+    height: 100vh;
+    position: relative;
+    font-family: 'Cabin', sans-serif;
+    color: #ffeded;
+    text-transform: uppercase;
+    font-size: 7vmin;
+    padding-left: 10%;
+    padding-right: 10%;
+}
+
+section:nth-child(odd)
+{
+    justify-content: flex-end;
 }
